requirements.txt
bcrypt==3.1.4
botocore==1.9.12
cffi==1.11.5
colorama==0.3.7
Django==1.10
django-debug-toolbar==1.9.1
docutils==0.14
jmespath==0.9.3
pyasn1==0.4.2
pycparser==2.18
python-dateutil==2.6.1
rsa==3.4.2
six==1.11.0
sqlparse==0.2.4
